+++ themes/theme-rui/src/components/Calendar.styles.ts
@@ -2,7 +2,7 @@ import { ThemeComponent, cva } from '@marigold/system';
 
 export const Calendar: ThemeComponent<'Calendar'> = {
   calendar: cva(
-    'rounded-lg border group-[trigger]/popover:border-0 border-border p-2'
+    'rounded-lg border group-[[data-trigger]]/popover:border-0 border-border p-2'
   ),
   calendarCell: cva([
     'relative flex size-9 items-center justify-center whitespace-nowrap rounded-lg',
@@ -10,13 +10,13 @@ export const Calendar: ThemeComponent<'Calendar'> = {
     'outline-offset-2 duration-150 [transition-property:color,background-color,border-radius,box-shadow]',
     'disabled:pointer-events-none data-[unavailable]:pointer-events-none data-focus-visible:z-10 data-[hovered]:bg-hover selected:bg-brand data-[hovered]:text-foreground selected:text-brand-foreground data-[unavailable]:line-through',
     'disabled:opacity-30 data-[unavailable]:opacity-30',
-    'focus:outline-0 focus-visible:outline-2 focus-visible:outline-ring/70',
+    'mixin-ring-focus-visible-borderless',
   ]),
   calendarControllers: cva([
     'size-9 rounded-lg',
     'text-muted-foreground/80',
-    'outline-offset-2 transition-colors',
-    'focus-visible:outline-2 focus-visible:outline-ring/70',
+    'transition-colors',
+    'mixin-ring-focus-visible',
   ]),
   calendarHeader: cva([
     'size-9 rounded-lg p-0 text-xs font-medium text-muted-foreground/80',