+++ themes/theme-rui/src/components/Input.styles.ts
@@ -6,28 +6,22 @@ export const inputContainer =
 export const inputDisabled =
   'disabled:cursor-not-allowed disabled:text-disabled-foreground disabled:bg-disabled';
 export const inputInvalid =
-  'group-error/field:border-destructive group-error/field:focus:border-destructive group-error/field:focus:ring-destructive/20';
-export const inputFocus = {
-  focus:
-    'focus:border-ring focus:outline-none focus:ring-[3px] focus:ring-ring/20',
-  'focus-within':
-    'focus-within:border-ring focus-within:outline-none focus-within:ring-[3px] focus-within:ring-ring/20',
-};
+  'group-invalid/field:border-destructive group-invalid/field:focus:border-destructive group-invalid/field:focus:ring-destructive/20';
 
-export const inputReadOnly = 'group-[readonly]/field:bg-muted';
+export const inputReadOnly = 'group-read-only/field:bg-muted';
 
 export const Input: ThemeComponent<'Input'> = {
   input: cva([
     inputContainer,
     inputDisabled,
     inputInvalid,
-    inputFocus['focus'],
+    'mixin-ring-focus-visible',
     inputReadOnly,
     'h-input',
     'placeholder:text-placeholder',
     '[&[type=file]]:border-solid [&[type=file]]:bg-background [&[type=file]]:p-0 [&[type=file]]:pr-3 [&[type=file]]:italic [&[type=file]]:text-muted-foreground/70 file:me-3 file:h-full file:border-0 file:border-r file:border-solid file:border-input file:bg-transparent file:px-3 file:text-sm file:font-medium file:not-italic file:text-foreground',
-    'group-[icon]/input:pl-8',
-    'group-[action]/input:pr-8',
+    'group-[[data-icon]]/input:pl-8',
+    'group-[[data-action]]/input:pr-7',
   ]),
   icon: cva([
     'pointer-events-none left-1',