+++ themes/theme-rui/src/components/Select.styles.ts
@@ -1,10 +1,5 @@
 import { type ThemeComponent, cva } from '@marigold/system';
-import {
-  inputContainer,
-  inputDisabled,
-  inputFocus,
-  inputInvalid,
-} from './Input.styles';
+import { inputContainer, inputDisabled, inputInvalid } from './Input.styles';
 
 export const Select: ThemeComponent<'Select'> = {
   icon: cva('text-muted-foreground/80'),
@@ -12,7 +7,7 @@ export const Select: ThemeComponent<'Select'> = {
     inputContainer,
     inputInvalid,
     inputDisabled,
-    inputFocus['focus'],
+    'mixin-ring-focus-visible',
     'h-input',
     'cursor-pointer',
     '*:data-placeholder:text-placeholder',